добавил плавную прокрутку
