+++ src/components/main-menu.pug
@@ -2,7 +2,7 @@ mixin main-menu(mod)
     nav(class="main-menu main-menu--" + mod)
         ul.main-menu__list
             li.main-menu__item
-                a(href="#welcome-section").main-menu__link Главная
+                a(href="#main-header").main-menu__link Главная
             li.main-menu__item
                 a(href="#about-section").main-menu__link Обо мне
             li.main-menu__item
