+++ src/components/welcome.pug
@@ -1,4 +1,4 @@
-section.welcome#welcome-section
+section.welcome
     .container.welcome__wrapper
         .welcome__content
             h1.welcome__title Личный сайт веб&nbsp;разработчика
