+++ src/index.pug
@@ -9,7 +9,7 @@ html(lang="ru")
     meta(name="Description", content="Разработка сайтов")
     title Portfolio | main page
   body
-    .wrapper
+    .wrapper#main-header
       .maincontent
         include components/parallax-mountain
         include components/header
