+++ src/main.js
@@ -5,7 +5,6 @@ if (process.env.NODE_ENV === "development") {
 
 import "./scripts/skills";
 import "./scripts/reviews-slider";
-import "./scripts/header";
 import "./scripts/parallax-mountain";
 import "./scripts/parallax-budda";
 import "./scripts/button-toggle";
