+++ src/styles/blocks/popap-menu.pcss
@@ -6,7 +6,7 @@
   background-color: #fff;
   width: 320px;
   height: 100vh;
-  display: none;
+  display: block;
   
   @include desktop {
     display: block;
