+++ src/styles/blocks/skils.pcss
@@ -21,7 +21,7 @@
     padding-left: 0;
     padding-right: 0;
     padding-top: 15px;
-    max-width: 375px;
+    justify-content: space-around;
   }
 }
